requirements.txt
accelerate==1.13.0
aiofile==3.9.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appdirs==1.4.4
arxiv==2.4.1
arxiv-mcp-server==0.4.11
asttokens==3.0.1
asyncer==0.0.8
attrs==26.1.0
Authlib==1.6.9
babel==2.18.0
beartype==0.22.9
beautifulsoup4==4.14.3
black==26.3.1
bleach==6.3.0
brotli==1.2.0
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.7
click==8.3.2
click-log==0.4.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
courlan==1.3.2
Crawl4AI==0.8.6
cryptography==46.0.7
cssselect==1.4.0
cyclopts==4.10.1
dataclasses-json==0.6.7
datasets==4.8.4
dateparser==1.4.0
decorator==5.2.1
deepdiff==9.0.0
defusedxml==0.7.1
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docling==2.84.0
docling-core==2.71.0
docling-ibm-models==3.13.0
docling-parse==5.7.0
docstring_parser==0.17.0
docutils==0.22.4
dspy==3.1.3
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fake-useragent==2.2.0
Faker==40.12.0
fastmcp==3.2.0
fastuuid==0.14.0
feedparser==6.0.12
filelock==3.25.2
filetype==1.2.0
firecrawl-py==4.22.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gepa==0.0.26
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.0
humanfriendly==10.0
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
instructor==1.15.1
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json_repair==0.58.7
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
jusText==3.0.2
keyring==25.7.0
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
lark==1.3.1
latex2mathml==3.79.0
litellm==1.83.0
lxml==5.4.0
lxml_html_clean==0.4.4
magika==0.6.3
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
markdownify==1.2.2
markitdown==0.1.5
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mcp==1.27.0
mcp-server-sqlite==2025.4.25
mdurl==0.1.2
model2vec==0.8.1
more-itertools==11.0.1
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nbformat>=5.0,<6.0
nltk==3.9.4
numpy==2.4.4
obsidiantools==0.11.0
omegaconf==2.3.0
onnxruntime==1.20.1
openai==2.30.0
openapi-pydantic==0.5.1
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
optuna==4.8.0
orderly-set==5.5.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
parso==0.8.6
patchright==1.58.2
pathable==0.5.0
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.4
playwright==1.58.0
playwright-stealth==2.0.3
pluggy==1.6.0
polyfactory==3.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pure_eval==0.2.3
py-key-value-aio==0.4.4
pyarrow==23.0.1
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.20.0
PyJWT==2.12.1
pylatexenc==2.10
pymdown-extensions==10.21.2
PyMuPDF==1.27.2.2
pyOpenSSL==26.0.0
pypdfium2==5.6.0
pyperclip==1.11.0
pyphen==0.17.2
pyreadline3==3.5.4
pytest==9.0.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-frontmatter==1.1.0
python-multipart==0.0.22
python-pptx==1.0.2
pytokens==0.4.1
pytz==2026.1.post1
pyvis==0.3.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
ragas==0.4.3
rank-bm25==0.2.2
rapidocr==3.7.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
rtree==1.4.1
ruff==0.15.9
safetensors==0.7.0
scikit-network==0.33.5
scipy==1.17.1
semchunk==3.2.5
setuptools==81.0.0
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlite-vec==0.1.9
sse-starlette==3.3.4
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
textstat==0.7.13
tiktoken==0.12.0
tld==0.13.2
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
trafilatura==2.0.0
traitlets==5.14.3
transformers==5.5.0
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
trimesh==4.11.5
typer==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uncalled-for==0.2.0
unclecode-litellm==1.81.13
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.43.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
yt-dlp==2026.3.17
zipp==3.23.0
zstandard==0.25.0
